# wasi: improves fd_write perf when not discarded