```bash
$ go test -run='^$' -bench '^Benchmark_fdWrite$' -count=5 > old.txt
$ go test -run='^$' -bench '^Benchmark_fdWrite$' -count=5 > new.txt
$ benchstat old.txt new.txt
name                    old time/op  new time/op  delta
_fdWrite/io.Writer-16    138ns ± 1%   132ns ± 3%  -3.95%  (p=0.008 n=5+5)
_fdWrite/io.Discard-16   137ns ± 6%   139ns ± 4%    ~     (p=0.421 n=5+5)
```